+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 # Jetpack-Compose-preview-creator Changelog
 
-
 ## [Unreleased]
+
+## [1.3.0] - 2024-12-16
 
 ### Added
 
@@ -25,6 +26,7 @@
 
 - Added settings for: parameter names, default values and theme
 
-[Unreleased]: https://github.com/EarlOfEgo/Jetpack-Compose-preview-creator/compare/v1.2.0...HEAD
+[Unreleased]: https://github.com/EarlOfEgo/Jetpack-Compose-preview-creator/compare/v1.3.0...HEAD
+[1.3.0]: https://github.com/EarlOfEgo/Jetpack-Compose-preview-creator/compare/v1.2.0...v1.3.0
+[1.2.0]: https://github.com/EarlOfEgo/Jetpack-Compose-preview-creator/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/EarlOfEgo/Jetpack-Compose-preview-creator/commits/v1.1.0
-[1.2.0]: https://github.com/EarlOfEgo/Jetpack-Compose-preview-creator/commits/v1.2.0